global import of jenar 4.3.2 +ar + acl

+++ jena-db/jena-tdb2/src/main/java/org/apache/jena/tdb2/solver/stats/StatsCollector.java
@@ -30,7 +30,7 @@ public class StatsCollector extends StatsCollectorBase<Node> {
     }
 
     @Override
-    protected Map<Node, Long> convert(Map<Node, Long> map) {
+    protected Map<Node, Integer> convert(Map<Node, Integer> map) {
         return map;
     }
 }
