+++ jena-db/jena-tdb2/src/main/java/org/apache/jena/tdb2/solver/stats/StatsCollectorNodeId.java
@@ -44,9 +44,9 @@ public class StatsCollectorNodeId extends StatsCollectorBase<NodeId> {
     }
 
     @Override
-    protected Map<Node, Long> convert(Map<NodeId, Long> stats) {
+    protected Map<Node, Integer> convert(Map<NodeId, Integer> stats) {
         // Predicate -> Count
-        Map<Node, Long> statsNodes = new HashMap<>(1000);
+        Map<Node, Integer> statsNodes = new HashMap<>(1000);
         for ( NodeId p : stats.keySet() ) {
             Node n = nodeTable.getNodeForNodeId(p);
             statsNodes.put(n, stats.get(p));
